requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.1
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
numpy==1.26.3
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
pillow==10.2.0
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
selenium==4.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.25
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
undetected-chromedriver==3.5.4
urllib3==2.1.0
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
gunicorn==20.1.0
